Scoring system inputs file to organize inputs to centralized location
