+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/settings/GamepadSettings.java
@@ -1,0 +1,16 @@
+package org.firstinspires.ftc.teamcode.settings;
+
+import com.aimrobotics.aimlib.gamepad.AIMPad;
+
+public class GamepadSettings {
+
+    // GAMEPAD 1
+    public static final double GP1_STICK_DEADZONE = 0.05;
+    public static final double GP1_TRIGGER_DEADZONE = 0.25;
+
+    public static final int EXPONENT_MODIFIER = 2;
+
+    // GAMEPAD 2
+    public static final double GP2_STICK_DEADZONE = 0.05;
+    public static final double GP2_TRIGGER_DEADZONE = 0.25;
+}
