+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/settings/ScoringSystemInputs.java
@@ -1,0 +1,16 @@
+package org.firstinspires.ftc.teamcode.settings;
+
+import com.aimrobotics.aimlib.gamepad.AIMPad;
+
+public class ScoringSystemInputs {
+
+    public boolean ADVANCE_AUTOMATION;
+
+    public ScoringSystemInputs(AIMPad aimPad, AIMPad aimPad2) {
+        updateInputs(aimPad, aimPad2);
+    }
+
+    public void updateInputs(AIMPad aimPad, AIMPad aimPad2) {
+        ADVANCE_AUTOMATION = aimPad.isAPressed();
+    }
+}
