+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/Outtake.java
@@ -6,8 +6,7 @@ import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.hardware.Servo;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.R;
-import org.firstinspires.ftc.teamcode.subsystems.settings.ConfigurationInfo;
+import org.firstinspires.ftc.teamcode.settings.ConfigurationInfo;
 
 public class Outtake extends Mechanism {
