+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/OuttakeSystem.java
@@ -6,8 +6,8 @@ import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.subsystems.settings.ConfigurationInfo;
-import org.firstinspires.ftc.teamcode.subsystems.settings.GamepadSettings;
+import org.firstinspires.ftc.teamcode.settings.ConfigurationInfo;
+import org.firstinspires.ftc.teamcode.settings.GamepadSettings;
 
 public class OuttakeSystem extends Mechanism {
     Outtake outtake;
