--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/settings/GamepadSettings.java
@@ -1,10 +1,0 @@
-package org.firstinspires.ftc.teamcode.subsystems.settings;
-
-public class GamepadSettings {
-
-    // GAMEPAD 1
-    public static final double GP1_STICK_DEADZONE = 0.05;
-    public static final double GP1_TRIGGER_DEADZONE = 0.25;
-
-    public static final int EXPONENT_MODIFIER = 2;
-}
