+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/InputModification.java
@@ -1,6 +1,6 @@
 package org.firstinspires.ftc.teamcode.util;
 
-import org.firstinspires.ftc.teamcode.subsystems.settings.GamepadSettings;
+import org.firstinspires.ftc.teamcode.settings.GamepadSettings;
 
 public class InputModification {
     /**
